+++ BLL/Models/OrderInfoViewModel.cs
@@ -13,8 +13,8 @@ namespace BLL.Models
         public string BookName { get; set; }
         public string AuthorName { get; set; }
         public int BookImageId { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public int Count { get; set; }
-        public double Cost { get; set; }
+        public decimal Cost { get; set; }
     }
 }
